Move IRCBot class to LibreNMS namespace.
